requirements.txt
discord.py
yfinance
alpha-vantage